requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
httpx==0.27.2
python-dateutil==2.9.0.post0
feedparser==6.0.11
rank-bm25==0.2.2
PyYAML==6.0.2
typer[all]==0.12.5
Jinja2==3.1.4
psycopg2-binary==2.9.10
loguru==0.7.2
aiosqlite==0.21.0
requests==2.32.5
greenlet==3.2.4